bootloader: Use "fast programming" on STM32
This is a procedure that flashes a full flash row (256 bytes), keeping
the flash voltage enabled during the entire row, rather than just one
doubleword (8 bytes) at a time. This makes the flashing a little faster
(worst case 22.4ms instead of 36.6ms for a full 2kbyte page), though
erasing a page still takes 40ms, so that dominates the total processing
time now.

+++ bootloader.cpp
@@ -93,7 +93,7 @@ static uint8_t commitToFlash(uint16_t address, uint16_t len) {
 
 	uint16_t offset = 0;
 	while (len > 0) {
-		uint8_t pageLen = len < FLASH_WRITE_SIZE ? len : FLASH_WRITE_SIZE;
+		uint16_t pageLen = len < FLASH_WRITE_SIZE ? len : FLASH_WRITE_SIZE;
 		uint8_t err = SelfProgram::writePage(address + offset, &writeBuffer[offset], pageLen);
 		if (err)
 			return err;
